+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/feat/DependencyPinningBundle.kt
@@ -3,7 +3,7 @@ package fluxo.conf.feat
 import fluxo.conf.FluxoKmpConfContext
 import fluxo.conf.data.VersionCatalogConstants.VC_PINNED_BUNDLE_ALIAS
 import fluxo.conf.impl.kotlin.KOTLIN_2_1
-import fluxo.conf.impl.kotlin.KOTLIN_PLUGIN_VERSION_RAW
+import fluxo.conf.impl.kotlin.KOTLIN_PLUGIN_VERSION_STRING
 import fluxo.conf.impl.logDependency
 import fluxo.log.d
 import fluxo.log.l
@@ -29,12 +29,11 @@ internal fun FluxoKmpConfContext.prepareDependencyPinningBundle() {
         // starting from Kotlin 2.1.0
         // Other plugins can bring incompatible versions of the compiler.
         // https://kotlinlang.slack.com/archives/C0KLZSCHF/p1729256644747559?thread_ts=1729151089.194689&cid=C0KLZSCHF
-        val version = runCatching { KOTLIN_PLUGIN_VERSION_RAW }.getOrNull()
-            ?: kotlinPluginVersion.toString()
         val compilerEmbeddable = object : ModuleIdentifier {
             override fun getGroup() = "org.jetbrains.kotlin"
             override fun getName() = "kotlin-compiler-embeddable"
         }
+        val version = KOTLIN_PLUGIN_VERSION_STRING
         pinnedDeps[compilerEmbeddable] = Pair(version, "Pinned to Kotlin plugin version")
     }